+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.1 under development
 
+- New #12: Add `RequestHeaderProvider` class that provides convenient access to request headers (@uzdevid)
 - Enh #17: Get request from provider into `RequestCookies` every time `get()` and `has()` methods are called (@vjik)
 
 ## 1.1.0 October 28, 2024